Bugfixes for sample project.

+++ SMWebRequest.m
@@ -301,7 +301,7 @@ static BOOL was_dealloced = NO;
 	NSInteger status = [response isKindOfClass:[NSHTTPURLResponse class]] ? [(NSHTTPURLResponse *)response statusCode] : 200;
 	
 	if (conn && response && (status < 200 || (status >= 300 && status != 304))) {
-		NSLog(@"Failed with HTTP status code %ld while loading %@", status, self);
+		NSLog(@"Failed with HTTP status code %i while loading %@", (int)status, self);
         
         SMErrorResponse *error = [[[SMErrorResponse alloc] init] autorelease];
         error.response = (NSHTTPURLResponse *)response;
